Draft muon veto channels support in Slow Control interface
And make timestamps default to datetime.now() where relevant
